refactor(chainAccountKit): move to own file
