+++ packages/orchestration/src/examples/stakeAtom.contract.js
@@ -70,14 +70,14 @@ export const start = async (zcf, privateArgs, baggage) => {
     'StakeAtom',
     M.interface('StakeAtomI', {
       makeAccount: M.callWhen().returns(M.remotable('ChainAccount')),
-      makeMakeAccountInvitation: M.call().returns(M.promise()),
+      makeAcountInvitationMaker: M.call().returns(M.promise()),
     }),
     {
       async makeAccount() {
         trace('makeAccount');
         return makeAccount().then(({ account }) => account);
       },
-      makeMakeAccountInvitation() {
+      makeAcountInvitationMaker() {
         trace('makeCreateAccountInvitation');
         return zcf.makeInvitation(
           async seat => {
